requirements.txt
flask
openai
gspread
google-auth
python-dotenv